Add final report and update README
