+++ crues_pi/crues_run
@@ -1,0 +1,45 @@
+#!/usr/bin/env bash
+
+if [[ $# -eq 0 ]]
+then
+    echo "Enter robot names as arguments."
+    exit 1
+else
+    echo "Checking execution permissions for ROS scripts..."
+    find ros_pkgs/ -name "*.py" -print0 | xargs -r0 chmod +x
+    for NAME in "$@"
+    do
+        if [[ ${NAME} = "blinky" ]]
+        then
+            IP=192.168.1.2
+        elif [[ ${NAME} = "inky" ]]
+        then
+            IP=192.168.1.3
+        elif [[ ${NAME} = "clyde" ]]
+        then
+            IP=192.168.1.4
+        else
+            echo "${NAME} is not a valid robot name."
+            exit 1
+        fi
+        read -p "Enter password for ${NAME}: " -s PWD
+        STATUS=""
+        while [[ ${STATUS} != "ok" ]]
+        do
+            if [[ ${PWD} = "" ]]
+            then
+                STATUS=none
+            else
+                STATUS=$(sshpass -p ${PWD} ssh crues@${IP} echo ok 2>&1)
+            fi
+            case ${STATUS} in
+                ok) ;;
+                none) echo ""; read -p "No password entered, try again: " -s PWD ;;
+                *"Permission denied"*) echo ""; read -p "Permission denied, try again: " -s PWD ;;
+                *) echo ""; echo ${STATUS}; exit 1 ;;
+            esac
+        done
+        echo ""
+        sshpass -p ${PWD} ssh crues@${IP} "roslaunch crues_control roomba.launch"
+    done
+fi
